Perform recursive lookups for group members that are themselves groups
In Google groups it is possible to make groups members of other groups. If A is
a member of group G and group G is a member of group H, then A should also be
considered a member of group H.

This change will add append any groups found when listing group members to the
array of groups to iterate through, so that eventually all top-level and nested
groups will be used to check membership for the user. It is not clear whether
cyclic group membership is permitted by Google but it is being checked for just
in case.