Redirect to uri even if --skip-provider-button
The normal behavior is that if a user is going to page
https://hostname/x but is not authenticated, they are sent to login and
then redirected back to https://hostname/x.

I discovered that if the flag --skip-provider-button is given, then
after login the user ends up at https://hostname/.

I believe this to be a bug. That behavior seems to happen because the
current URI is only extracted when redirecting to the sign-in page form,
not when the OAuth flow is started, so if that form is skipped, the
current URL is lost. Therefore, I made a change to extract the current
page to send into the OAuth flow.

This change fixes that bug and adds a test.